Types

FastBEAST.KMeansTreeOptions — Type

KMeansTreeOptions <: TreeOptions

Is the datatype that describes which tree the create_tree function creates.

Fields

iterations: number of iterations on each level, default is one iterations

nchildren: defines the number of children of each node, default is two

nmin: defines the minimum amount of datapoints which are needed in a cluster so that it gets split up in subclusters, default is 1

maxlevel: defines the maximum amount of levels, default is 100.

algorithm: defines which algorithm is used. The :naive approach is not recommended. Default is the wrapped ParallelKMeans algorithm.

FastBEAST.KMeansTreeNode — Type

KMeansTreeNode{T} <: AbstractNode

Is the datatype of each node in the K-Means Clustering Tree.

Fields

parent::Union{KMeansTreeNode{T},Nothing}: is the superordinate cluster of of the represented cluster

children::Union{Vector{KMeansTreeNode{T}}, Nothing}: all directly subordinated clusters of the represented cluster

level::Integer: the level of the represented cluster

center: the center by which the cluster is defined

radius: the euclidian distance between the center and the farthest away point

data::T: array containig the indices of the points in this cluster

Functions

FastBEAST.create_tree — Function

create_tree(points::Array{SVector{D, T}, 1}; treeoptions)

Creates an algebraic tree for an givn set of datapoints. The returned datastructure is the foundation for the algorithms in this package.

Arguments

points::Array{SVector{D, T}, 1}: is an array of SVector. Each SVector contains in general two or three float values, which discribe the position in the space.

Keywords

treeoptions::TreeOptions: this keyword defines by which tree is build. TreeOptions is an abstract type which either can be BoxTreeOptions or KMeansTreeOptions. Default type is BoxTreeOptions.

Returns

AbstractNode: the root of the created tree. AbstractNode is an abstract type which either can be BoxTreeNode or KMeansTreeNode, depending on the keyword.

FastBEAST.iscompressable — Function

iscompressable(sourcenode::AbstractNode, testnode::AbstractNode)

Determins whether two nodes of a tree are comressable. The criteria differs between the Box Tree and the K-Means Clustering Tree. For the K-Means Clustering Tree two nodes can be compressed, if the distance between the centers of two clusters is greater than the sum of their radius multiplied by a factor of 1.5. For the Box Tree two nodes can be compressed, if the distance between the centers of the two boxes is greater than the sum of the distances of each box's center to one of its corners multiplied by a factor of 1.1.

Arguments

sourcenode::AbstractNode: the node which is observed

testnode::AbstractNode: the node which is tested for compression

Returns

true: if the input nodes are compressable

false: if the input nodes are not compressable